chore: verify credential refactor
