+++ packages/vc.js/src/checkCredential/checkDate.ts
@@ -1,13 +1,5 @@
 import moment from "moment";
 
-// thanks Mozilla... https://bugzilla.mozilla.org/attachment.cgi?id=351506&action=diff
-// Killer regex to parse RFC3339 dates
-export const RFC_3339 = new RegExp(
-  "^([0-9]{4})-([0-9]{2})-([0-9]{2})" +
-    "([Tt]([0-9]{2}):([0-9]{2}):([0-9]{2})(.[0-9]+)?)?" +
-    "(([Zz]|([+-])([0-9]{2}):([0-9]{2})))?"
-);
-
 // Tweaked to allow for leap seconds
 export const ISO_8601_FULL = /^([\+-]?\d{4}(?!\d{2}\b))((-?)((0[1-9]|1[0-2])(\3([12]\d|0[1-9]|3[01]))?|W([0-4]\d|5[0-2])(-?[1-7])?|(00[1-9]|0[1-9]\d|[12]\d{2}|3([0-5]\d|6[1-6])))([T]((([01]\d|2[0-3])((:?)[0-5]\d)?|24\:?00)([\.,]\d+(?!:))?)?(\17[0-6]\d([\.,]\d+)?)?([zZ]|([\+-])([01]\d|2[0-3]):?([0-5]\d)?)?)?)?$/;
 
@@ -35,13 +27,10 @@ export const checkDate = (
     valid: false,
     warnings: []
   };
+  // open source vc, make sure validation tests date strings are the right format for the date
 
-  if (!ISO_8601_FULL.test(datetime)) {
+  if (!ISO_8601_FULL.test(datetime)) {  
     res.warnings.push(`${datetime} is not a legal ISO 8601 Date Time.`);
-  }
-
-  if (!RFC_3339.test(datetime)) {
-    res.warnings.push(`${datetime} is not a legal RFC 3339 Date Time.`);
   }
 
   moment.suppressDeprecationWarnings = true;
